Version 0.30.3
- Feature #133 : Warning when an update is available
- Feature: Warning when MHX is not installed or enabled.
- Bug #137: Hair custom properties were not displayed correctly in configuration mode if MustardUI Naming Convention was disabled.
- Bug: Fixed some issues related to the Diffeomorphic support (thanks to @mokujinh). The add-on should now correctly find the Diffeomorphic version and not show version issues if the Diffeomorphic support is not requested
- Bug: Fixed an issue that prevented the morphs to be used in some models (Diffeomorphic version not supported error).
- Bug: Warning added on Outfits Switcher due to Blender bug (https://projects.blender.org/blender/blender/issues/116061). It will be reverted as soon as the bug is fixed.

Note: if you are using the Armature Outfit Switcher, please remove BOTH the outfit piece and the collection (in this order!), and disable it. Since this feature might causes crashes, I hidden it with the Experimental Features addon flag: thus enable it to remove this feature from your models! You can find this flag in the addon settings (Blender Settings -> search for MustardUI) similarly to the Developer and Debug modes.